oeeBackend/requirements.txt
astroid==1.6.4
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.1
pytz==2018.4
six==1.11.0
wrapt==1.10.11
